fix: remove `serialize-javascript` from client bundle
